[FIX] account_financial_risk: Display the Financial Risk page correctly when the contact is a person but does not have a parent_id.
During migration to V17, the conversion from attribute to Python expression was incorrect. As a result, this page became visible when a contact is a person but has a parent_id. The correct logic should display the field only when the contact is a person and does not have a parent_id.