Revert "[IMP] account_credit_control: Send communication emails in background"
This reverts commit 834e7ade44908ef4a1c3d456b80b4be8330a708b.

Commit:
https://github.com/odoo/odoo/commit/543a3ad204232cc67ae9e099f43960b664e8de19

Contains the comment that
```
```

So a thread is generally assumed to take at most 2 cursors.
Because of this, it is essentially a very bad idea to open an unlimited number of cursors.
(On a decently sized DB, about 400 lines would be processed in parallel...)

Because things are launched in their own cursor, this doesn't impact the main request,
so the run is marked as done and all lines are marked as queued,
despite the emails never being scheduled.

There is already a way to manage this kind of cases in the OCA that is called `queue_job`.
Therefore this is the module that should handle this.

If for some reason one would still want the threaded code nonetheless,
it should be put in a custom project module, as it relies on some environment hypothesis
that can't be tested from within the code.

Note that one test is split into two; although it is better to avoid code duplication,
the test was cheating the setup, and therefore not representative of the real user flow is
when users click on the interface buttons.